deliver.sh

+++ scripts/deliver.sh
@@ -1,3 +1,2 @@
-#if pgrep hotpoll; then pkill -f hotpoll; fi
-pkill -f hotpoll
-JENKINS_NODE_COOKIE=dontKillMe nohup java -jar $1  &
+if pgrep -f hotpoll; then pkill -f hotpoll; fi
+JENKINS_NODE_COOKIE=dontKillMe nohup java -jar ${1} &
